requirements.txt
google-cloud-storage 
google-cloud-bigquery 
python-dotenv
pyarrow
google-api-core==2.11.0
google-api-python-client==2.81.0
google-auth==2.16.2
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.58.0
oauth2client==4.1.3
oauthlib==3.2.2
httplib2==0.21.0
pandas==1.5.3
json5==0.9.11
jsonpointer==2.3
jsonschema==4.17.3
db-dtypes
httplib2
argparse
google-ads
glom